auth_helper: Add option to specify client_secret
Without this adaption, I weren't able to connect to the API in the
"production" environment. I also weren't able to find a constructor
argument to set this secret.

Of course, connection to the production env also requires setting the
following variables to correct values:

    auth_helper.CLIENT_ID,
    auth_helper.CLIENT_SECRET,
    auth_helper.DEVELOPER_TOKEN,
    auth_helper.ENVIRONMENT.

It also requires putting one's refresh token into the file specified by

    auth_helper.REFRESH_TOKEN.